+++ packages/vue/src/devCheck.ts
@@ -1,5 +1,6 @@
 if (__BROWSER__ && __DEV__) {
-  console[(console as any).info ? 'info' : 'log'](
+  // @ts-ignore `console.info` cannot be null error
+  console[console.info ? 'info' : 'log'](
     `You are running a development build of Vue.\n` +
       `Make sure to use the production build (*.prod.js) when deploying for production.`
   )